+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes qsf-integration project will be documented in this file.
 - implement tracking tags for search query
 - add exception to the log message in the ElasticFilter
 - add tags to search query
+- map nested fields
 
 ### Changed
 - sort facet by key in case of all facets are loaded